+++ pkg/controller/linstorcontroller/linstorcontroller_controller_test.go
@@ -36,21 +36,10 @@ func TestNewConfigMapForPCS(t *testing.T) {
 					Namespace: "default-ns",
 				},
 				Data: map[string]string{
-					"linstor.toml": `[config]
-
-[debug]
-
-[log]
-
-[db]
+					"linstor.toml": `[db]
   connection_url = "etcd://etcd.svc:5000/"
-  [db.etcd]
 
-[http]
-
-[https]
-
-[ldap]
+[logging]
 `,
 					"linstor-client.conf": `[global]
 controllers = http://test.default-ns.svc:3370
@@ -78,22 +67,11 @@ controllers = http://test.default-ns.svc:3370
 					Namespace: "default-ns",
 				},
 				Data: map[string]string{
-					"linstor.toml": `[config]
-
-[debug]
-
-[log]
-
-[db]
+					"linstor.toml": `[db]
   connection_url = "etcd://secure.etcd.svc:443/"
   ca_certificate = "/etc/linstor/certs/ca.pem"
-  [db.etcd]
 
-[http]
-
-[https]
-
-[ldap]
+[logging]
 `,
 					"linstor-client.conf": `[global]
 controllers = http://test.default-ns.svc:3370
@@ -122,24 +100,13 @@ controllers = http://test.default-ns.svc:3370
 					Namespace: "default-ns",
 				},
 				Data: map[string]string{
-					"linstor.toml": `[config]
-
-[debug]
-
-[log]
-
-[db]
+					"linstor.toml": `[db]
   connection_url = "etcd://secure.etcd.svc:443/"
   ca_certificate = "/etc/linstor/certs/ca.pem"
   client_certificate = "/etc/linstor/certs/client.cert"
   client_key_pkcs8_pem = "/etc/linstor/certs/client.key"
-  [db.etcd]
 
-[http]
-
-[https]
-
-[ldap]
+[logging]
 `,
 					"linstor-client.conf": `[global]
 controllers = http://test.default-ns.svc:3370
@@ -149,7 +116,7 @@ controllers = http://test.default-ns.svc:3370
 			},
 		},
 		{
-			name: "with-https-auth",
+			name: "with-https-auth-and-log-level",
 			spec: &piraeusv1.LinstorController{
 				ObjectMeta: metav1.ObjectMeta{
 					Name:      "test",
@@ -162,6 +129,7 @@ controllers = http://test.default-ns.svc:3370
 					LinstorClientConfig: shared.LinstorClientConfig{
 						LinstorHttpsClientSecret: "secret",
 					},
+					LogLevel: shared.LogLevelTrace,
 				},
 			},
 			expected: &corev1.ConfigMap{
@@ -170,26 +138,18 @@ controllers = http://test.default-ns.svc:3370
 					Namespace: "default-ns",
 				},
 				Data: map[string]string{
-					"linstor.toml": `[config]
-
-[debug]
-
-[log]
-
-[db]
-  connection_url = "etcd://etcd.svc:5000/"
-  [db.etcd]
-
-[http]
-
-[https]
+					"linstor.toml": `[https]
   enabled = true
   keystore = "/etc/linstor/https/keystore.jks"
   keystore_password = "linstor"
   truststore = "/etc/linstor/https/truststore.jks"
   truststore_password = "linstor"
 
-[ldap]
+[db]
+  connection_url = "etcd://etcd.svc:5000/"
+
+[logging]
+  linstor_level = "TRACE"
 `,
 					"linstor-client.conf": `[global]
 controllers = https://test.default-ns.svc:3371